security connector refactor
