+++ grpc.gemspec
@@ -223,11 +223,14 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/security/credentials/oauth2/oauth2_credentials.h )
   s.files += %w( src/core/lib/security/credentials/plugin/plugin_credentials.h )
   s.files += %w( src/core/lib/security/credentials/ssl/ssl_credentials.h )
-  s.files += %w( src/core/lib/security/security_connector/alts_security_connector.h )
+  s.files += %w( src/core/lib/security/security_connector/alts/alts_security_connector.h )
+  s.files += %w( src/core/lib/security/security_connector/fake/fake_security_connector.h )
   s.files += %w( src/core/lib/security/security_connector/load_system_roots.h )
   s.files += %w( src/core/lib/security/security_connector/load_system_roots_linux.h )
-  s.files += %w( src/core/lib/security/security_connector/local_security_connector.h )
+  s.files += %w( src/core/lib/security/security_connector/local/local_security_connector.h )
   s.files += %w( src/core/lib/security/security_connector/security_connector.h )
+  s.files += %w( src/core/lib/security/security_connector/ssl/ssl_security_connector.h )
+  s.files += %w( src/core/lib/security/security_connector/ssl_utils.h )
   s.files += %w( src/core/lib/security/transport/auth_filters.h )
   s.files += %w( src/core/lib/security/transport/secure_endpoint.h )
   s.files += %w( src/core/lib/security/transport/security_handshaker.h )
@@ -651,11 +654,14 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/security/credentials/oauth2/oauth2_credentials.cc )
   s.files += %w( src/core/lib/security/credentials/plugin/plugin_credentials.cc )
   s.files += %w( src/core/lib/security/credentials/ssl/ssl_credentials.cc )
-  s.files += %w( src/core/lib/security/security_connector/alts_security_connector.cc )
+  s.files += %w( src/core/lib/security/security_connector/alts/alts_security_connector.cc )
+  s.files += %w( src/core/lib/security/security_connector/fake/fake_security_connector.cc )
   s.files += %w( src/core/lib/security/security_connector/load_system_roots_fallback.cc )
   s.files += %w( src/core/lib/security/security_connector/load_system_roots_linux.cc )
-  s.files += %w( src/core/lib/security/security_connector/local_security_connector.cc )
+  s.files += %w( src/core/lib/security/security_connector/local/local_security_connector.cc )
   s.files += %w( src/core/lib/security/security_connector/security_connector.cc )
+  s.files += %w( src/core/lib/security/security_connector/ssl/ssl_security_connector.cc )
+  s.files += %w( src/core/lib/security/security_connector/ssl_utils.cc )
   s.files += %w( src/core/lib/security/transport/client_auth_filter.cc )
   s.files += %w( src/core/lib/security/transport/secure_endpoint.cc )
   s.files += %w( src/core/lib/security/transport/security_handshaker.cc )
