+++ grpc.gyp
@@ -472,11 +472,14 @@
         'src/core/lib/security/credentials/oauth2/oauth2_credentials.cc',
         'src/core/lib/security/credentials/plugin/plugin_credentials.cc',
         'src/core/lib/security/credentials/ssl/ssl_credentials.cc',
-        'src/core/lib/security/security_connector/alts_security_connector.cc',
+        'src/core/lib/security/security_connector/alts/alts_security_connector.cc',
+        'src/core/lib/security/security_connector/fake/fake_security_connector.cc',
         'src/core/lib/security/security_connector/load_system_roots_fallback.cc',
         'src/core/lib/security/security_connector/load_system_roots_linux.cc',
-        'src/core/lib/security/security_connector/local_security_connector.cc',
+        'src/core/lib/security/security_connector/local/local_security_connector.cc',
         'src/core/lib/security/security_connector/security_connector.cc',
+        'src/core/lib/security/security_connector/ssl/ssl_security_connector.cc',
+        'src/core/lib/security/security_connector/ssl_utils.cc',
         'src/core/lib/security/transport/client_auth_filter.cc',
         'src/core/lib/security/transport/secure_endpoint.cc',
         'src/core/lib/security/transport/security_handshaker.cc',
